requirements.txt
amqp==5.1.1
annotated-types==0.5.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.3.0
async-timeout==4.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
boto3==1.28.17
botocore==1.31.17
Brotli==1.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colour==0.1.5
contourpy==1.1.0
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==41.0.3
cssselect2==0.7.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==4.2
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-crontab==0.7.1
django-debug-toolbar==4.1.0
django-excel==0.0.10
django-extensions==3.2.3
django-import-export==3.2.0
django-indian-numbers==0.0.2
django-js-asset==2.1.0
django-qr-code==3.1.1
django-qrcode==0.3
django-redis==5.3.0
django-storages==1.13.2
django-tables2==2.6.0
django-tenants==3.5.0
django-tenants-celery-beat==0.2.0
django-timezone-field==5.1
django-weasyprint==2.2.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.2
fonttools==4.42.1
glcontext==2.4.0
html5lib==1.1
idna==3.4
ipython==8.15.0
isosurfaces==0.1.0
jedi==0.19.0
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.1
lml==0.1.0
lxml==4.9.3
mailtrap==2.0.1
ManimPango==0.4.3
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
MarkupPy==1.14
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
moderngl==5.8.2
# moderngl-window==2.4.4
mpmath==1.3.0
multipledispatch==1.0.0
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
# Pillow==10.0.0
pipenv==2023.6.26
platformdirs==3.8.0
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydub==0.25.1
pydyf==0.8.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyglet==2.0.9
Pygments==2.16.1
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyOpenGL==3.1.7
pyparsing==3.0.9
pypdf==3.14.0
pyperclip==1.8.2
pyphen==0.14.0
pypng==0.20220715.0
pyrr==0.10.3
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
redis==4.6.0
reportlab==3.6.13
requests==2.31.0
rich==13.5.2
s3transfer==0.6.1
scipy==1.11.1
screeninfo==0.8.1
segno==1.5.2
sentry-sdk==1.29.2
six==1.16.0
skia-pathops==0.8.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
svgelements==1.9.6
svglib==1.5.1
sympy==1.12
tablib==3.5.0
tenant-schemas-celery==2.2.0
texttable==1.7.0
thermocouples-reference==0.20
tinycss2==1.2.1
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.16
validators==0.22.0
vine==5.0.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
weasyprint==60.1
webencodings==0.5.1
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3
# The conflict is caused by:
#     The user requested Pillow==10.0.0
#     matplotlib 3.7.2 depends on pillow>=6.2.0
#     moderngl-window 2.4.4 depends on Pillow<10 and >=9
